+++ src/authentication/authentication.service.ts
@@ -124,11 +124,4 @@ export class AuthenticationService {
       data: { email: newEmail },
     });
   }
-
-  async updateUsername(userId: number, newUsername: string) {
-    return await this.prisma.user.update({
-      where: { id: userId },
-      data: { username: newUsername },
-    });
-  }
 }
